+++ scripts/restart.sh
@@ -19,7 +19,7 @@ function schedule_restart() {
     fi
 
     for ((counter=15; counter>=1; counter--)); do
-        if [[ -n $RCON_ENABLED ]] && [[ "${RCON_ENABLED,,}" == "true" ]]; then
+        if [[ -n $RESTAPI_ENABLED ]] && [[ "${RESTAPI_ENABLED,,}" == "true" ]]; then
             if check_is_server_empty; then
                 ew ">>> Server is empty, restarting now"
                 if [[ -n $WEBHOOK_ENABLED ]] && [[ "${WEBHOOK_ENABLED,,}" == "true" ]]; then
@@ -29,7 +29,7 @@ function schedule_restart() {
             else
                 ew ">>> Server has still players"
             fi
-            rconcli broadcast "$(get_time) AUTOMATIC RESTART IN $counter MINUTES"
+            restapi_announce "$(get_time) AUTOMATIC RESTART IN $counter MINUTES"
         fi
         if [[ -n $RESTART_DEBUG_OVERRIDE ]] && [[ "${RESTART_DEBUG_OVERRIDE,,}" == "true" ]]; then
             sleep 1
@@ -38,13 +38,13 @@ function schedule_restart() {
         fi
     done
 
-    if [[ -n $RCON_ENABLED ]] && [[ "${RCON_ENABLED,,}" == "true" ]]; then
-        rconcli broadcast "$(get_time) Saving world before restart..."
-        rconcli save
-        rconcli broadcast "$(get_time) Saving done"
+    if [[ -n $RESTAPI_ENABLED ]] && [[ "${RESTAPI_ENABLED,,}" == "true" ]]; then
+        restapi_announce "$(get_time) Saving world before restart..."
+        restapi_save
+        restapi_announce "$(get_time) Saving done"
         sleep 15
         kill -SIGTERM "${PLAYER_DETECTION_PID}"
-        rcon "Shutdown 10"
+        restapi_shutdown 10 "$(get_time) Server restarting..."
         if [[ -n $WEBHOOK_ENABLED ]] && [[ "${WEBHOOK_ENABLED,,}" == "true" ]]; then
             send_stop_notification
         fi